refactor: simplify auth flow
