+++ apps/api/src/app/auth/community.auth.module.config.ts
@@ -1,10 +1,9 @@
 import { MiddlewareConsumer, ModuleMetadata, Provider, RequestMethod } from '@nestjs/common';
-import { JwtModule } from '@nestjs/jwt';
+import { JwtModule, JwtService } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import passport from 'passport';
-
 import { AuthProviderEnum, PassportStrategyEnum } from '@novu/shared';
-
+import { CommunityUserRepository, CommunityOrganizationRepository, CommunityMemberRepository } from '@novu/dal';
 import { JwtStrategy } from './services/passport/jwt.strategy';
 import { AuthController } from './auth.controller';
 import { UserModule } from '../user/user.module';
@@ -16,9 +15,10 @@ import { EnvironmentsModuleV1 } from '../environments-v1/environments-v1.module'
 import { JwtSubscriberStrategy } from './services/passport/subscriber-jwt.strategy';
 import { RootEnvironmentGuard } from './framework/root-environment-guard.service';
 import { ApiKeyStrategy } from './services/passport/apikey.strategy';
-import { injectCommunityAuthProviders } from './inject-auth-providers';
 import { AuthService } from './services/auth.service';
 import { RolesGuard } from './framework/roles.guard';
+import { CommunityAuthService } from './services/community.auth.service';
+import { CommunityUserAuthGuard } from './framework/community.user.auth.guard';
 
 const AUTH_STRATEGIES: Provider[] = [JwtStrategy, ApiKeyStrategy, JwtSubscriberStrategy];
 
@@ -27,31 +27,49 @@ if (process.env.GITHUB_OAUTH_CLIENT_ID) {
 }
 
 export function getCommunityAuthModuleConfig(): ModuleMetadata {
+  const baseImports = [
+    PassportModule.register({
+      defaultStrategy: PassportStrategyEnum.JWT,
+    }),
+    JwtModule.register({
+      secret: process.env.JWT_SECRET,
+      signOptions: {
+        expiresIn: 360000,
+      },
+    }),
+  ];
+
+  // Start with essential providers only
+  const baseProviders = [...AUTH_STRATEGIES, AuthService, RolesGuard, RootEnvironmentGuard, JwtService];
+
+  // Wherever is the string token used, override it with the provider
+  const injectableProviders = [
+    {
+      provide: 'USER_REPOSITORY',
+      useClass: CommunityUserRepository,
+    },
+    {
+      provide: 'ORGANIZATION_REPOSITORY',
+      useClass: CommunityOrganizationRepository,
+    },
+    {
+      provide: 'MEMBER_REPOSITORY',
+      useClass: CommunityMemberRepository,
+    },
+    {
+      provide: 'AUTH_SERVICE',
+      useClass: CommunityAuthService,
+    },
+    {
+      provide: 'USER_AUTH_GUARD',
+      useClass: CommunityUserAuthGuard,
+    },
+  ];
+
   return {
-    imports: [
-      OrganizationModule,
-      SharedModule,
-      UserModule,
-      PassportModule.register({
-        defaultStrategy: PassportStrategyEnum.JWT,
-      }),
-      JwtModule.register({
-        secret: process.env.JWT_SECRET,
-        signOptions: {
-          expiresIn: 360000,
-        },
-      }),
-      EnvironmentsModuleV1,
-    ],
+    imports: [...baseImports, EnvironmentsModuleV1, SharedModule, UserModule, OrganizationModule],
     controllers: [AuthController],
-    providers: [
-      ...USE_CASES,
-      ...AUTH_STRATEGIES,
-      ...injectCommunityAuthProviders({ repositoriesOnly: false }),
-      AuthService,
-      RolesGuard,
-      RootEnvironmentGuard,
-    ],
+    providers: [...baseProviders, ...injectableProviders, ...USE_CASES],
     exports: [
       RolesGuard,
       RootEnvironmentGuard,
@@ -61,7 +79,6 @@ export function getCommunityAuthModuleConfig(): ModuleMetadata {
       'USER_REPOSITORY',
       'MEMBER_REPOSITORY',
       'ORGANIZATION_REPOSITORY',
-      ...USE_CASES,
     ],
   };
 }
