+++ apps/api/src/app/shared/shared.module.ts
@@ -24,6 +24,9 @@ import {
   TopicSubscribersRepository,
   UserRepository,
   WorkflowOverrideRepository,
+  CommunityUserRepository,
+  CommunityMemberRepository,
+  CommunityOrganizationRepository,
 } from '@novu/dal';
 import {
   analyticsService,
@@ -47,7 +50,6 @@ import {
 
 import { isClerkEnabled, JobTopicNameEnum } from '@novu/shared';
 import { JwtModule } from '@nestjs/jwt';
-import { injectCommunityAuthProviders } from '../auth/inject-auth-providers';
 import packageJson from '../../../package.json';
 
 function getDynamicAuthProviders() {
@@ -56,7 +58,22 @@ function getDynamicAuthProviders() {
 
     return eeAuthPackage.injectEEAuthProviders();
   } else {
-    return injectCommunityAuthProviders();
+    const userRepositoryProvider = {
+      provide: 'USER_REPOSITORY',
+      useClass: CommunityUserRepository,
+    };
+
+    const memberRepositoryProvider = {
+      provide: 'MEMBER_REPOSITORY',
+      useClass: CommunityMemberRepository,
+    };
+
+    const organizationRepositoryProvider = {
+      provide: 'ORGANIZATION_REPOSITORY',
+      useClass: CommunityOrganizationRepository,
+    };
+
+    return [userRepositoryProvider, memberRepositoryProvider, organizationRepositoryProvider];
   }
 }
