+++ apps/api/src/app/user/usecases/create-user/create-user.usecase.ts
@@ -9,8 +9,8 @@ export class CreateUser {
   async execute(data: CreateUserCommand): Promise<UserEntity> {
     const user = new UserEntity();
 
-    user.email = data.email ? data.email.toLowerCase() : null;
-    user.firstName = data.firstName ? data.firstName.toLowerCase() : null;
+    user.email = data.email ? data.email.toLowerCase() : '';
+    user.firstName = data.firstName ? data.firstName.toLowerCase() : '';
     user.lastName = data.lastName ? data.lastName.toLowerCase() : data.lastName;
     user.profilePicture = data.picture;
     user.showOnBoarding = true;
