+++ apps/api/src/app/user/usecases/index.ts
@@ -1,4 +1,4 @@
-import { CreateUser } from '@novu/application-generic';
+import { CreateUser } from './create-user/create-user.usecase';
 import { GetMyProfileUsecase } from './get-my-profile/get-my-profile.usecase';
 import { UpdateOnBoardingUsecase } from './update-on-boarding/update-on-boarding.usecase';
 import { UpdateOnBoardingTourUsecase } from './update-on-boarding-tour/update-on-boarding-tour.usecase';
